updated the ccd model for case record
