+++ src/main/java/uk/gov/hmcts/reform/hmc/api/services/HearingsDataServiceImpl.java
@@ -4,7 +4,6 @@ import static uk.gov.hmcts.reform.hmc.api.utils.Constants.C100;
 import static uk.gov.hmcts.reform.hmc.api.utils.Constants.CASE_TYPE_OF_APPLICATION;
 import static uk.gov.hmcts.reform.hmc.api.utils.Constants.FL401;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
@@ -26,12 +25,6 @@ import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.hmc.api.mapper.FisHmcObjectMapper;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.CaseDetailResponse;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.Element;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.Flags;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.PartyDetails;
-import uk.gov.hmcts.reform.hmc.api.model.ccd.flagdata.FlagDetail;
 import uk.gov.hmcts.reform.hmc.api.model.request.HearingValues;
 import uk.gov.hmcts.reform.hmc.api.model.response.CaseCategories;
 import uk.gov.hmcts.reform.hmc.api.model.response.CaseFlags;
@@ -53,14 +46,14 @@ import uk.gov.hmcts.reform.hmc.api.utils.Constants;
 @SuppressWarnings("unchecked")
 public class HearingsDataServiceImpl implements HearingsDataService {
 
-    public static final String EMPTYSTRING = " ";
-
     @Value("${ccd.ui.url}")
     private String ccdBaseUrl;
 
     @Autowired CaseApiService caseApiService;
 
     @Autowired private ResourceLoader resourceLoader;
+
+    @Autowired CaseFlagDataServiceImpl caseFlagDataService;
 
     protected static final List<String> HEARING_CHANNELS =
             Arrays.asList("INTER", "TEL", "VID", "ONPPRS");
@@ -105,7 +98,7 @@ public class HearingsDataServiceImpl implements HearingsDataService {
                         + Constants.UNDERSCORE
                         + caseDetails.getData().get(Constants.APPLICANT_CASE_NAME);
         String caseSlaStartDateMapper = (String) caseDetails.getData().get(Constants.ISSUE_DATE);
-        getCcdCaseData(caseDetails);
+        caseFlagDataService.getCcdCaseData(caseDetails);
         JSONObject screenFlowJson = null;
         JSONParser parser = new JSONParser();
         Resource resource = resourceLoader.getResource("classpath:ScreenFlow.json");
@@ -226,97 +219,5 @@ public class HearingsDataServiceImpl implements HearingsDataService {
     public List<HearingLinkData> getHearingLinkData(
             HearingValues hearingValues, String authorisation, String serviceAuthorization) {
         return new ArrayList<>();
-    }
-
-    public CaseDetailResponse getCcdCaseData(CaseDetails caseDetails) throws IOException {
-        ObjectMapper objectMapper = FisHmcObjectMapper.getObjectMapper();
-
-        CaseDetailResponse ccdResponse =
-                objectMapper.convertValue(caseDetails, CaseDetailResponse.class);
-
-        return ccdResponse;
-    }
-
-    public void getCaseFlagData(ServiceHearingValues hearingsData, CaseDetails caseDetails)
-            throws IOException {
-
-        List<PartyFlagsModel> partiesFlagslList = new ArrayList<>();
-        List<PartyDetailsModel> partyDetailsModelList = new ArrayList<>();
-        CaseDetailResponse ccdResponse = getCcdCaseData(caseDetails);
-        List<Element<PartyDetails>> applicantLst = ccdResponse.getCaseData().getApplicants();
-
-        addPartyFlagData(partiesFlagslList, partyDetailsModelList, applicantLst);
-
-        List<Element<PartyDetails>> respondedLst = ccdResponse.getCaseData().getRespondents();
-
-        addPartyFlagData(partiesFlagslList, partyDetailsModelList, respondedLst);
-
-        CaseFlags caseFlags = CaseFlags.caseFlagsWith().flags(partiesFlagslList).build();
-
-        hearingsData.setCaseFlags(caseFlags);
-
-        hearingsData.setParties(partyDetailsModelList);
-    }
-
-    private void addPartyFlagData(
-            List<PartyFlagsModel> partiesFlagslList,
-            List<PartyDetailsModel> partyDetailsModelList,
-            List<Element<PartyDetails>> applicantLst) {
-        IndividualDetailsModel individualDetailsModel;
-        String uuid;
-        PartyDetails partyDetails;
-        PartyDetailsModel partyDetailsModel;
-        for (Element<PartyDetails> party : applicantLst) {
-            uuid = party.getId().toString();
-            partyDetails = party.getValue();
-            partiesFlagslList.addAll(getPartyFlagsModel(partyDetails, uuid));
-
-            individualDetailsModel =
-                    IndividualDetailsModel.individualDetailsWith()
-                            .firstName(partyDetails.getFirstName())
-                            .lastName(partyDetails.getLastName())
-                            .build();
-
-            partyDetailsModel =
-                    PartyDetailsModel.partyDetailsWith()
-                            .partyID(uuid)
-                            .partyName(
-                                    partyDetails.getFirstName()
-                                            + EMPTYSTRING
-                                            + partyDetails.getLastName())
-                            .partyType(PartyType.IND)
-                            .partyRole(Constants.APPLICANT)
-                            .individualDetails(individualDetailsModel)
-                            .build();
-
-            partyDetailsModelList.add(partyDetailsModel);
-        }
-    }
-
-    private List<PartyFlagsModel> getPartyFlagsModel(PartyDetails partyDetails, String uuid) {
-        PartyFlagsModel partyFlagsModel = null;
-        List<PartyFlagsModel> partyFlagsModelList = new ArrayList<>();
-        Flags flag = partyDetails.getPartyLevelFlag();
-        List<Element<FlagDetail>> detailsLST = flag.getDetails();
-
-        for (Element<FlagDetail> flagDetailElement : detailsLST) {
-            FlagDetail flagDetail = flagDetailElement.getValue();
-
-            partyFlagsModel =
-                    PartyFlagsModel.partyFlagsModelWith()
-                            .partyId(uuid)
-                            .partyName(
-                                    partyDetails.getFirstName()
-                                            + EMPTYSTRING
-                                            + partyDetails.getLastName())
-                            .flagId(flagDetail.getData().getFlagCode())
-                            .flagStatus(flagDetail.getData().getStatus())
-                            .flagParentId("")
-                            .flagDescription(flagDetail.getData().getFlagComment())
-                            .build();
-            partyFlagsModelList.add(partyFlagsModel);
-        }
-
-        return partyFlagsModelList;
     }
 }
