+++ src/main/java/uk/gov/hmcts/reform/hmc/api/utils/Constants.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.hmc.api.utils;
 
 public final class Constants {
 
+    public static final String EMPTYSTRING = " ";
     public static final Boolean TRUE = true;
     public static final Boolean FALSE = false;
     public static final String CASE_TYPE_OF_APPLICATION = "caseTypeOfApplication";
